1) В классе SearchActivity ключи для передачи данных вынесены в константу. 2) В классе SharedPreferenceRepository параметры передаются в конструктор через private val. 3) Из класса SettingsInteractorImpl убрана зависимость от SettingsManager. 4) В класса TracksRepositoryImpl из сети получается модель трека TrackDto и выполняется маппинг TrackDto в Track. 5) Класс TracksSearchResponse удален. 6) В классе App записана ссылка на Application и используется везде, где нужен контекст. 7) Удален весь закоментированный код. 8) Исправлена проблема: если включить темную тему в настройках, то при каждом заходе на экран настроек свитч "Тёмная тема" будет в неактивном состоянии и чтобы выключить темную тему нужно два раза на него нажать 9) Исправлена проблема: все данные, передаваемые между слоями Data и Domain, должны проходить через DTO - TrackDTO не используется, используется одна и та же модель Track для работы с треками во всех слоях 10) Исправлена проблема: Код должен быть хорошо структурирован, без избыточных зависимостей между слоями - Имеется избыточная зависимость SettingsInteractorImpl от SettingsManager из data слоя. Интерактор может зависеть только от репозиториев

+++ app/src/main/java/com/practicum/playlistmaker/data/network/TracksRepositoryImpl.kt
@@ -17,12 +17,11 @@ class TracksRepositoryImpl(
         val response = networkClient.doRequest(TracksSearchRequest(expression))
         if (response.resultCode == 200) {
             return (response as TrackResponse).results.map {
-//                Track(it.trackName, it.artistName, it.trackTime, it.artworkUrl100, it.collectionName,it.releaseDate,it.primaryGenreName,it.country,it.previewUrl)  }
                 Track(
-                    it.trackId,
+                    it.trackId.toInt(),
                     it.trackName,
                     it.artistName,
-                    it.trackTimeMillis,
+                    it.trackTimeMillis.toInt(),
                     it.artworkUrl100,
                     it.collectionName,
                     it.releaseDate,
